requirements/development.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.3.0
distlib==0.3.6
Django==4.2.4
django-filter==23.2
django-nose==1.4.7
djangorestframework==3.14.0
drf-spectacular==0.26.4
filelock==3.12.2
Flask==2.3.2
Flask-Celery-Helper==1.1.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-redis==0.4.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.1
Markdown==3.4.4
MarkupSafe==2.1.3
nose==1.3.7
numpy==1.25.1
pandas==2.0.3
platformdirs==3.7.0
prompt-toolkit==3.0.39
py4j==0.10.9.7
pyspark==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.2
rpds-py==0.9.2
six==1.16.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
virtualenv==20.23.1
wcwidth==0.2.6
Werkzeug==2.3.6
